Add a Route53 subdoman for the website for easy access
Add a data source to import the existing Route 53 zone
Alias the ELB to a sub domain name.  Per the documentation, this is the preferred method over using a CNAME since the zone is being hosted in Route 53.  Due to the way AWS handles Aliases on the backend, resolution is faster and it is a free service whereas CNAMEs are charged.

Remove the output from the terraform script as the ELB is now aliased to a known domain name
Remove the bucket_name variable as it was a leftover from a previous iteration of this lab project
Cleanup some of the formatting and add additional comments to ease readability

Also, tested with Terraform 0.12.2 and terraform plan threw some errors - it looks like the expected format has been modified slightly.  The aws provider hasn’t been updated to work with .12.2 yet anyway, so this issue will be addressed later on.