Minimum reconfig needed to make the autoscaling group work
Remove the aws_instance resource since the auto scaling group will create the instances
Add a launch config resource and the autoscaling group resource
Change the target group attachment from an instance attachment to an autoscaling attachment

The config still needs to be cleaned up and validated that it works like expected.  Currently, the provisioning script doesn’t create the webpage so this will need to be updated so its compatible with how the auto scaling group handles provisioning.  There are also no scaling policies setup yet, so they also need to be created.