Fix the bootstrap script so it runs on boot
Changed the Launch Config user_data parameter to use the bootstrap shell script (vs the inline script) - it can accept either a cloud-init script or a regular shell script per the documentation.
Modified the shell script to run as root as it was originally designed to be run as EC2-User with the remote-exec provisioner when it was part of the aws_instance resource.
Launch Configs cannot be changed after creation, so any changes cause a destroy and recreate.  For this reason the TF recommends setting create_before_destroy in the lifecycle block AND not specifying the name parameter, use a name_prefix instead.
Update some tags and formatting.