Add registration and login forms to the flask application to allow user input
The Flask extension wtforms has some classes that help build forms.  Add a new forms.py file that sets up 2 new form classes - these classes are converted to html when run.
Add new routes to the main file for the Registration and Login pages.  Add some methods to these routes to handle the POST method
Add 2 new templates for the Registration and Login pages.  The HTML in the tutorial uses some bootstrap style sheet stuff, however I am trying to omit all of that to keep it as simple as possible for now.
Add data validation checks and feedback to indicate why a validation fails - this is handled inside of the template pages (register and login)
Add a flash message to indicate when a user account has been created successfully - right now this just means the user filled out the registration page correctly
Add a secret key to help secure the forms from cross-site request forgery attacks.  For now this key is part of the code, but later needs to be changed to something more secure since posting the key to GitHub makes this insecure…
Add a simple header to ease navigation among the pages